Initial multithreaded support for *nix.
Well, for FreeBSD and NetBSD... OpenDoors has some crazy requirements
that you can't meet with pthreads (and shouldn't with the Windows API)

Microsoft has this to say:
TerminateThread is a dangerous function that should only be used in
the most extreme cases. You should call TerminateThread only if you
know exactly what the target thread is doing, and you control all of
the code that the target thread could possibly be running at the time
of the termination.

Obviously "threads is hard" is a most extreme case.