Fix RIP autodetection
When ESC! is sent, rip replies with the form "RIPSCRIP015410", but
OpenDoors was only detecting the "RIP" part, then clearing the key
buffer and hoping for the best. If any of the trailing bytes were
received "later", they would end up in the input queue.

RIPTerm in DOSBox with a telnet modem would split the string in
arbitrary ways, ensuring that parts of it were, in fact, received
later.

For extra hilarity, an 'S' in the key buffer would stop a file being
displayed early, and a 'P' would pause it (with no prompt). One of
the two would randomly happen.

Now, it still enables RIP if "RIP" is detected, but the discards
up to 11 more bytes in the same 660ms timeout that's being used for
RIP detection in the first place.